# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: EasySale POS API
  description: |
    EasySale is a white-label, configuration-driven point-of-sale solution designed to adapt to any retail business type.
    
    ## Authentication
    The API uses two authentication methods:
    - **Cookie Authentication**: httpOnly cookie (`auth_token`) set after successful login
    - **Bearer Token**: JWT token in Authorization header for API clients
    
    ## Multi-Tenant Architecture
    All requests are scoped to a tenant. The tenant context is extracted from the authenticated user's JWT token.
    
    ## Offline-First Design
    The system supports offline operation with local SQLite storage and background synchronization.
  version: 1.0.0
  contact:
    name: EasySale Support
  license:
    name: Proprietary
servers:
  - url: http://localhost:7945
    description: Development server
  - url: https://api.easysale.local
    description: Production server (LAN)
...
security:
  - cookieAuth: []
  - bearerAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: httpOnly cookie containing JWT token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token in Authorization header
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        message:
          type: string
          description: Detailed error message
        code:
          type: string
          description: Error code
      required:
        - error
...
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          description: Username
          example: admin
        password:
          type: string
          format: password
          description: Password
          example: password123
      required:
        - username
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT token (also set as httpOnly cookie)
        user:
          $ref: '#/components/schemas/UserResponse'
        expires_at:
          type: string
          format: date-time
          description: Token expiration time
      required:
        - token
        - user
        - expires_at
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        display_name:
          type: string
          nullable: true
        role:
          type: string
          enum: [admin, manager, cashier, specialist, inventory_clerk, technician]
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true
        store_id:
          type: string
          nullable: true
        station_policy:
          type: string
          enum: [any, specific, none]
        station_id:
          type: string
          nullable: true
        permissions:
          type: array
          items:
            type: string
      required:
        - id
        - tenant_id
        - username
        - email
        - role
        - station_policy
        - permissions
...
paths:
...
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login
      description: Authenticate user and receive JWT token (also set as httpOnly cookie)
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
              description: httpOnly cookie containing auth token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Account requires store/station assignment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many login attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...